requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
alembic>=1.7.0
redis>=4.0.0
celery>=5.2.0
pydantic>=1.8.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.8.0
websockets>=10.0
googlemaps>=4.6.0
python-dotenv>=0.19.0
requests>=2.26.0
kafka-python>=2.0.2
confluent-kafka>=2.1.0  # More robust Kafka client
bcrypt>=4.0.0 
watchdog>=6.0.0
mysql-connector-python>=8.0.0
